moved the history logic into its own module from which the model now inherits so that it can access all the data regarding the history of calculations
