+++ src/constants/constants.js
@@ -1,3 +1,3 @@
-export const MAX_OPERAND_LENGTH = 9;
+export const MAX_OPERAND_LENGTH = 12;
 export const MAX_HISTORY_LENGTH = 5;
 export const ID_LENGTH = 8;
