+++ src/model/index.js
@@ -1,3 +1,2 @@
-import Model from './Model';
-
-export default Model;
+export { default as Model } from './Model';
+export { default as PubSub } from './PubSub';
